requirements.txt
requests
tensorflow
pytest
pylint
ipdb
jupyterlab
numpy
pandas
matplotlib
seaborn
scikit-learn
google-cloud-storage
